Adds Shiny app for sleeping site metadata
Implements an interactive Shiny application for managing and visualizing sleeping site metadata.

Includes features for uploading, editing, and downloading site data as a CSV file.

The app also displays the sites on a Leaflet map and allows users to select rows in a table to highlight corresponding markers on the map.